requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
ultralytics
opencv-python-headless
numpy<2.0
